Fix error rate computation always reporting 0. Change experiment configuration to use 1 repetition instead of 2. Created a new scene to allow executing the mouse test on the computer screen, so we can compare if all the math we do in this program results in similar throughput and metrics as the literature.

+++ Assets/Scripts/TestMeasurements.cs
@@ -75,7 +75,7 @@ public class TestMeasurements {
         results["totalTrials"] = totalTrials;
         results["totalMissedTargets"] = totalMissedTargets;
         results["totalMovementTime"] = totalMovementTime;
-        results["errorRate"] = totalMissedTargets / totalTrials;
+        results["errorRate"] = (float)totalMissedTargets / (float)totalTrials;
 
         float averageMovementTime = totalMovementTime / totalTrials;
         results["averageMovementTime"] = averageMovementTime;
